argscichat/requirements.txt
datasets==1.6.1
dill==0.3.3
gensim==4.0.1
huggingface-hub==0.0.8
Keras==2.4.3
Keras-Preprocessing==1.1.2
matplotlib==3.4.2
multiprocess==0.70.11.1
networkx==2.5.1
nltk==3.5
numpy==1.18.5
pandas==1.2.4
scikit-image==0.18.1
scikit-learn==0.24.0
scipy==1.4.1
simplejson==3.17.2
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
tensorflow-probability==0.11.1
tokenizers==0.10.2
tqdm==4.56.0
transformers==4.6.1
